Create README.md
White canes are going high-tech. Inventors have equipped white canes with ultrasonic devices that detect obstacles up to nine feet away. Vibrations in the cane’s handle warn users of potential hazards in their path. In this project, we are using ultrasonic sensor that detect obstacles up to 12 feet away and when the ultrasonic sensor detects an obstacle within 1.5 feet the buzzer beeps intermittently. When the obstacle gets even closer, within 1 foot the buzzer beeps continuously. For demonstration purpose, we are also printing the measured distance from an obstacle for every second in Tera Term.